Creates a database of all the shows, seasons and episodes in a folder
TV_DB_structure.py creates the initial DB using sqlite3 and os walk to
search through the folders and files where the episodes are kept.

TV_search.py searches through the DB and compares them to what TVDB
has.  It then informs the user on which episodes may be missing.